+++ src/clojure_lsp/server.clj
@@ -17,7 +17,8 @@
    (clojure_lsp
      ClojureExtensions
      ExtraMethods
-     CursorInfoParams)
+     CursorInfoParams
+     ClojuredocsParams)
    (java.util.concurrent CompletableFuture
                          CompletionException)
    (java.util.function Supplier)
@@ -160,7 +161,7 @@
 
   (^CompletableFuture rename [_ ^RenameParams params]
     (start :rename
-           (async-request params handlers/rename ::interop/workspace-edit)))
+           (async-request params handlers/rename ::interop/workspace-edit-or-error)))
 
   (^CompletableFuture hover [_ ^HoverParams params]
     (start :hover
@@ -246,7 +247,7 @@
 
   (^CompletableFuture linkedEditingRange [_ ^LinkedEditingRangeParams params]
     (start :linkedEditingRange
-           (async-request params handlers/linked-editing-ranges ::interop/linked-editing-ranges))))
+           (async-request params handlers/linked-editing-ranges ::interop/linked-editing-ranges-or-error))))
 
 (deftype LSPWorkspaceService []
   WorkspaceService
@@ -378,6 +379,11 @@
       (start :cursor-info-log
              (future
                (sync-notification params handlers/cursor-info-log))))
+
+    (^CompletableFuture clojuredocsRaw [^ClojuredocsParams params]
+      (start :clojuredocsRaw
+             (CompletableFuture/completedFuture
+               (sync-request params handlers/clojuredocs-raw ::interop/clojuredocs-raw))))
 
     (^CompletableFuture shutdown []
       (log/info "Shutting down")