+++ test/clojure_lsp/shared_test.clj
@@ -64,7 +64,13 @@
                                                                (h/file-path "/user/project/src/cljs")}}
                                     :project-root-uri (h/file-uri "file:///user/project")})
     (is (= "foo.bar"
-           (shared/uri->namespace (h/file-uri "file:///user/project/src/clj/foo/bar.clj") db/db)))))
+           (shared/uri->namespace (h/file-uri "file:///user/project/src/clj/foo/bar.clj") db/db))))
+  (testing "when an invalid source-path with a valid source-path prefixing it"
+    (swap! db/db medley/deep-merge {:settings {:source-paths #{(h/file-path "/user/project/src/clj")}}
+                                    :project-root-uri (h/file-uri "file:///user/project")})
+    (with-redefs [shared/directory? (constantly true)]
+      (is (= nil
+             (shared/uri->namespace (h/file-uri "file:///user/project/src/cljs/foo/bar.clj") db/db))))))
 
 (deftest conform-uri
   (testing "lower case drive letter and encode colons"